From Gerri: Public key files should always be unlinked at the end of use by rootd/proofd, but it is true that when for some reason the unlinking was unsuccessful, the next daemon may have run in the problems you observed.
I have re-written RpdSavePubKey in such a way that it returns
an error code, runs fchown before writing and checks any
possible failures.

Within RpdUpdateAuthTab, for RpdSavePubKey failures due
to already existing files which cannot be overwritten,
a dummy entry is entered in rpdauthtab so that a new offset
(so a new file name) is used.

I have also found a place in TAuthenticate where the value
of the returned offset was not checked for non-negativity,
which was probably the reason of the crash.


git-svn-id: http://root.cern.ch/svn/root/trunk@9764 27541ba8-7e3a-0410-8455-c3a389f83636