From Timur Pocheptsov
I've updated root from CVS, there is a resource leak in TGWin32GL.cxx. It
was fixed in my friday's files, but it is not fixed in CVS.
It's fixed in my new diff file.

And now we can make objects transparent by simply pressing 't' or 'T' key
- in material mode we can rotate scene with left button and make an object
transparent by right click.


git-svn-id: http://root.cern.ch/svn/root/trunk@9916 27541ba8-7e3a-0410-8455-c3a389f83636