From Federico Carminati   I have attached few mods necessary to have root working with the IBM compiler on the Mac. It now works through all the benchmarks, so you can recompile on G5 with optimisation and quote a performance finally.  The remaining changes I had to make were:
- Make DYLD_LIBRARY_PATH point to the right place (SHAME!)

- Avoid dynamic casting of null pointers. This is a phylosophical point
which can be debated. I see pro and cons to have a compiler that does not
like them, however I had to protect the code in a couple of places.

- Remove the invalid trap from main programs. I consider this almost a bug
of the compiler.

- Remove -flat-namespace from linking libraries. Again I think the
behaviour is highly debatable, but I know the linker should get better in
Tiger

- Add -DR__BROKEN_FUNCTION_TEMPLATES to the compilation

Please check the mods and put them in the head asap so they do not get out
of synch. Let me know about the performance on the G5, I am curious to see
whether we have gained anything!!!


git-svn-id: http://root.cern.ch/svn/root/trunk@10204 27541ba8-7e3a-0410-8455-c3a389f83636