From Axel Naumann: I ran into a tiny problem, trying to merge (actually hadd) labelled histos, where the labels were distributed differently (I was making use of FindBin(label), adding labels as needed).
The attached patch allows Merge to add histos with labels, where
same-name labels are summed, and new bins with new labels are added to
the histo as needed. So merging is not bin number centric but bin label
centric, if all histos have labels.

This label mode is chosen if all histos have at least one bin with a
label which is != "" or 0. A warning is printed if some of them have
labels, others don't; Merge the falls back to bin number merging.

I also found a tiny bug: with two histos being identical, all bin
contents are added, including the underflow (and the overflow,
accordingly). For that, FindBin(GetBinCenter(0)) is called. But that
rebins the histo, as it's in the underflow range. I protected that by
calling ResetBit(kCanRebin) and afterwards SetBit, if it was set before.
The line has the comment "// reset, otherwise getting the under/overflow
will rebin".

And I removed two unneeded local vars (ibin, bin).


git-svn-id: http://root.cern.ch/svn/root/trunk@10257 27541ba8-7e3a-0410-8455-c3a389f83636