From Philippe: On windows, library are required to be linked with all their dependencies. In particular this means that the cintdlls have to be linked against libCint.dll. Also the current ROOT build system has an intermediary rootcint executable which is used for creating the ROOT dictionary.  This intermediary executable is actually build statically (aka directly against the .o files).
The problem is that if this statically linked executable loads the cintdlls,
it also has to load the cintdlls.  This results in the presence in memory of
2 version of the CINT data structure ... Which lead to weird, wrong behavior
preventing the generation of the dictionary.

As far as I can tell, the makecint/rootcint phase works as well wethere the
cintdlls are present or not (at least for the ROOT dictionary).  Hence to
allow the ROOT Dictionary to be buildable even after the cintdlls have been
generated I added a protection to prevent their loading if we are on windows
AND we are running rootcint/makecint.


git-svn-id: http://root.cern.ch/svn/root/trunk@10265 27541ba8-7e3a-0410-8455-c3a389f83636