From Philippe: The crash for iobug.C(1,0) was fixed in a previous patch.
The bad results in iobug.C(1,1) and iobug.C(2,1) were a result of badly
formed files (i.e. the file produce do not contain the correct data!).

This patch resolves the solution by insuring that all the level of base
classes have their own TBranchElement object except when inside a split
collection.

When the base class or any of its own base class has no data member, there
is no branch created (this was already the case).

The TBranchElement corresponding to the base class itself does NOT have a
TLeafElement (but has at least one sub-TBranchElement).  Its fObject points
to the start of the sub-part of the object which correspond to its base
class.

In addition, this patch corrects TBranch::GetSubBranch to give the correct
result even if the hiearchy of branch is more than 2 deep.

This patch also correct GetValueClones and GetValueSTL to add the base class
offset only once.


git-svn-id: http://root.cern.ch/svn/root/trunk@10555 27541ba8-7e3a-0410-8455-c3a389f83636