From Gerri: The last minute change in the convention for the first argument of TCondition::TimedWait(), i.e. from this
   Int_t   TimedWait(ULong_t secs, ULong_t nanoSecs = 0);

to this

   Int_t   TimedWait(ULong_t secs, ULong_t nanoSecs);
   Int_t   TimedWait(ULong_t ms);

had some nasty effect on the behaviour of TXNetFile in
asynchronous mode, showing up dramatically only on SLC3 and
Solaris (for some reason not really understood, on RH73 and
MacOSx things continued to work ... and, needless to say, the
first check I did after the commit has been on these two machines).

The patch fixes the related problem in TXInputBuffer.cxx and
also an additional bug found by Fabrizio in TXNetConn.cxx just
before Xmas.
There are also an additional protection and a couple of cosmetic
changes in TXPhyConnection.


git-svn-id: http://root.cern.ch/svn/root/trunk@10870 27541ba8-7e3a-0410-8455-c3a389f83636