From Philippe, Fix for a problem reported by Jacek Holeczek
I finally understood and solved the problem.

First the loading was considered successful is any loaded library has the
name of the requested library embedded in its full pathname (which is the
wrong behavior and is fixed by the attached patch).
In addition any loaded files with the string '.a','.so','.sl' embedded in
their full pathname was considered a library (which is the wrong behavior
and is fixed by the attached patch).

Hence in you case:


/.automount/p2a00/root/exports/home/holeczek/test_subdir_name_must_contain_G
4_string/./x.cxx

Was considered a library (because of the second and third character) and was
considered to be equivalent to the library 'G4/G4' because of the 'G4'
string in the directory name.

The attached patch solves this problem and modify the search behavior for
existing library.
Now we have

Once a file named XXX.[a|so|sl|dl] has been loaded, gSystem->Load will
consider that file matching [/*/][lib]XXX.[a|so|sl|dl].

In other word you can not load 2 libraries whose full pathname differs only
by the pathname and/or the last extension.


git-svn-id: http://root.cern.ch/svn/root/trunk@10887 27541ba8-7e3a-0410-8455-c3a389f83636