From Philippe: This patch extends the notion of TTree variable aliases to allow them to be used as part of single data name.
For example, you can now do:
    tree->SetAlias("theGoodTrack","event.fTracks[3]");
    tree->Draw("theGoodTrack.fPx"); // same as "event.fTracks[3].fPx"

This patch includes a little clean-up in TTreeFormula::DefinedVariable
to better separate the different phases of the parsing (dimension requests,
leaf finding and exploration into the user class).

This patch also includes an update to DrawTest to test for the new alias
feature and in case of failure to print the corresponding Draw command.


git-svn-id: http://root.cern.ch/svn/root/trunk@11141 27541ba8-7e3a-0410-8455-c3a389f83636