From Philippe: On windows we encounter a fatal (but semi-random) problem which is directly due to the fact that the buffer used in Type.cxx (G__buf) to hold the result of G__type2string is under-sized (only G__ONELINE instead of G__LONGLINE as in G__type2string).
Also I (re)notice that there __no__ protection or warnings in case of buffer
over-run in G__type2string.  The consequence of this missing test is that a
user will discover that his class name are too long (and hence that he needs
to increase G__LONGLINE) only when he stumble upon some random crashes due
to the buffer run (assuming that he is patient enough to actually discover
that this is the problem).

I corrected the problem in Type.cxx AND I added a
test in G__type2string to print an error message if there is a buffer
over-run.  Alternatively,
we could actually terminate the program in the failing case.


git-svn-id: http://root.cern.ch/svn/root/trunk@11212 27541ba8-7e3a-0410-8455-c3a389f83636