From Richard Maunder: More fixes to the external viewer creation/closedown. We now:
i) Always call Modified() and Update() we create an external viewer 3D
(non-pad tyep - x3d or ogl) so that an externally spwaned viewer is
painted automcatically when created. For an internal TViewer3DPad we
asssume this was

ii) No longer paint the pad when the external viewer is closed. This
leaves the pad blank - but this is preferable to causing a double redraw
of the pad as the external viewer is closed - very slow....

This all needs to be revisited as part of supporting multiple viewers
properly on the pad.


git-svn-id: http://root.cern.ch/svn/root/trunk@11356 27541ba8-7e3a-0410-8455-c3a389f83636