From Markus,
Introduce a new class TDirectory::TContext to keep track and
restore the current directory.  With this construct C++
exceptions will be guaranteed to properly restore the
current directory pointer.

For example code like:
   TDirectory *sav = gDirectory;
   mydirectory->cd();
   if (...) {
      ....
      sav->cd();
      return;
   } else if (...) {
      ....
      sav->cd();
      return;
   }
   sav->cd;
   return;
can be replaced with the simplier and exception safe:
   TDirectory::TContext context(gDirectory,mydirectory);
   if (...) {
      ....
      return;
   } else if (...) {
      ....
      return;
   }
   return;

TDirectory::WriteTObject and TDirectory::WriteObjectAny now
store the object directly in 'this' directory (instead
of storing it in the (possibly unrelated) current directory).

TObject::Write has been modified accordingly (it contained the
main example of code relying on the previous behavior)

This patches correct a the problem demonstrated by
the test currentFileKey in roottest/root/io/interface


git-svn-id: http://root.cern.ch/svn/root/trunk@11826 27541ba8-7e3a-0410-8455-c3a389f83636