From Anna Kreshuk: Methods to compute Least trimmed squares regression coefficients were added to the TLinearFitter class.
Now robust linear fitting can be performed for the datasets, in which
outliers influence the ordinary least-squares fit. The maximal breakdown
point of the algorithm is |(floor((n-p)/2) +1)/n, which means that
LTS fit stays in a bounded region whenever floor((n-p)/2) or fewer
observations are replaced by arbitary points.

The implementation is based on the article "Computing LTS regression
for Large Data Sets" by Peter J. Rousseeuw and Katrien Van Driessen.


git-svn-id: http://root.cern.ch/svn/root/trunk@12136 27541ba8-7e3a-0410-8455-c3a389f83636