From Christian Holm: I've made a patch for some updates to the packaging stuff in ROOT.
The most important change is, that in Debian and Redhat packages
`libroot', I install the libraries in `/usr/lib/root/<version>' (where
version is something like 5.03), and I put symbolic links to the ROOT
libraries in `/usr/lib/root'

        /usr/lib/root/libTreePlayer.so.5.03 -> /usr/lib/root/5.03/libTreePlayer.so.5.03
        /usr/lib/root/libTreePlayer.so -> /usr/lib/root/5.03/libTreePlayer.so.5.03

And in `/etc/root/system.rootrc' we get

	Unix.Root.*.DynamicPath: .:/usr/lib/root/5.03

This allows users to only install `libroot' and not be forced to install
`libroot-dev' for the dynamic loading of ROOT plugins to work.

It also means, that users can more easily upgrade ROOT.  That is,
suppose you have `libroot' version 4.04.03 installed, and you have some
application that depends on that, say `mn-fit'.  Now, you want to
upgrade to the next production version, say 5.03.01, but `mn-fit' hasn't
been recompiled against the new ROOT libraries.   As the libraries in
`libroot' are all versioned (end in `.so.<version>'), the two version of
`libroot' can easily co-exist.  However, the libraries should load the
correct libraries for the particular version, which is ensured by the
line in `/etc/root/system.rootrc'.  The name of the package `libroot' is
changed to `libroot<major>' where <major> is the major version number of
ROOT (e.g., `5').

For the packages, root-rootd, and root-proofd, I removed the (X)Inetd
stuff, so that these services should be started by SYSV Init scripts
instead.

Also, I updated the `README/INSTALL' file to address some recent
complaints about the RPM building.  I also added section numbers and a
Table of Contents.

In `configure', I added the directory `/usr/include/postgresql/8.0' to
the search path for `lib-pq.h', as that's where the headers are put now
a days on Debian `sid'.

In `xrootd/Module.mk', I took out the `xrdpwdgsi' executable from the
`ALLEXECS' variable, as it doesn't seem to be build.

I added the man page `xrdpwdadmin.1'.


git-svn-id: http://root.cern.ch/svn/root/trunk@12478 27541ba8-7e3a-0410-8455-c3a389f83636