First version of TTreeSQL facility. This allow the storing and restoring of TTree into an SQL database
  TSQLServer *dbserver = TSQLServer::Connect("mysql://localhost:3306/rootDev","rootdevel", "r00tg6ys");
  dbserver->Query("drop table ntuple;");
  ntuple  = new TTreeSQL(dbserver, "rootDev", "ntuple");

If the database contain a table named ntuple, this will connect to it
and let you use any of TTree functionality on (Scan, Draw, etc.)

If the database does not contain a table named ntuple, it will created
once you called the first Fill on the TTree.

This version supports the leaflist type of branches.


git-svn-id: http://root.cern.ch/svn/root/trunk@12483 27541ba8-7e3a-0410-8455-c3a389f83636