from Andreas Peters: I just added a function, GetEndpointUrl(), which allows you to find out after redirection to which host you are redirected.
So, if you give a list of Alien URLs in a Dataset, the master can try to
open all files over the redirector and will end up at some hosts. It can
then replace the alien url with the url, where it could open the files.
This is cool, because you could set the connection time only to 2s and if
the master cannot open the file, it will just skip it in the dataset.
Or the master would trigger the staging over xroootd ... aso ...
so, if you could put this in soon, would be cool.


git-svn-id: http://root.cern.ch/svn/root/trunk@12650 27541ba8-7e3a-0410-8455-c3a389f83636