From Jeroen Hegeman: Last week I got really stuck looking at memory addresses and missing fBits when I got some help from Paul Russo. We found out that the core of the problem is pretty simple. In TStreamerBuffer::WriteBufferAux there is a bit of code that writes the fBits and if the object is referenced it also writes the pidf.
The problem is that no offset table is generated for this 'fBits' branch so
reading this branch will result in corruption. (You write variable-
size things and always read fixed-size things.)

The solution is to force (in TBranchElement's constructor) to support a
variable size in the case of fBits (case kBits).


git-svn-id: http://root.cern.ch/svn/root/trunk@12715 27541ba8-7e3a-0410-8455-c3a389f83636