From Gerri: This is the patch for Marek's problem.
Basically, ReinitSelector() does not need to be called if running in
synchronous mode (new argument to TProofPlayer::Finalize()), and the aclic
mode must be recorded and used when reinitialization is needed.

It remains the problem with ::GetSelector that I mentioned in the
previous mail, which affects, for example, cases when one runs once
with "<selec>.C+", modifies <selec>.C and runs again forgetting the
'+'.


git-svn-id: http://root.cern.ch/svn/root/trunk@12829 27541ba8-7e3a-0410-8455-c3a389f83636