From Paul Russo:
We extend the support for pointers to varying length arrays
of objects to include the cases:

  1) Objects whose classes do not have a ClassDef macro, but do
     have dictionaries loaded (no default Streamer, but we do have
     TStreamerInfo).

  2) Objects of classes with custom Streamer functions.

  3) Reading an object from a file when we do not have the
     dictionary loaded for the object's class, and then writing
     that object out to a file again (emulated objects).  This
     allows us to copy trees without have to have the class
     dictionaries loaded.

Note that we still do not support the case of STL objects.

There is a problem with the old implementation and the current
implementation in that memory is leaked in the case of reading
a varying length array of pointers to objects.  The memory for
the objects read on the previous read is leaked because the code
is written in such a way that we have already overwritten the
length of the old variable array with the length of the new one
before we get to the place where we must iterate through the
array and free the old objects.  The code must be changed to
take note of the old value of the varying length when the new
value is read, so that it may be used later to free the old
set of objects.


git-svn-id: http://root.cern.ch/svn/root/trunk@12991 27541ba8-7e3a-0410-8455-c3a389f83636