Add new includes in the matrix package to be used for forward declarations of some matrix classes that will be replaced by templated classes starting in January 2006. Whenever you have a forward declaration like   class TMatrixD; it should be replaced by   #include "TMatrixDfwd.h" and same for classes :
  TMatrixDBase,TMatrixDSparse,TMatrixDSym,TMatrixDUtils
  TMatrixFBase,TMatrixFSparse,TMatrixFSym,TMatrixFUtils
  TVectorD,TVectorF,TVector
In the current CVS version, the include file TMatrixDfwd.h
simply contains the forward declaration "class TmatrixD"
In the new version of the include that will be included in
the next ROOT version in January, the forward declaration
will be templated to match the templated version of the new
matrix package. No other changes will be required on the user side.

WE STRONGLY RECOMMEND USERS OF THE MATRIX PACKAGE TO ANTICIPATE
THIS CHANGE BY REPLACING THEIR FORWARD DECLARATION BY AN INCLUDE
LIKE "TMatrixDfwd.h". By doing so, you will have a smooth transition
between the two versions of ROOT.


git-svn-id: http://root.cern.ch/svn/root/trunk@13102 27541ba8-7e3a-0410-8455-c3a389f83636