From Christian Holm: Here's a patch and some files for ROOT, to fix some problems in the Debian build.
The file `root-plugin-minuit2.control' plus some changes in the existing
scripts, adds the package `root-plugin-minuit2' containing the new
Minuit implementation.   `root-bin.png' is used on Debian to integrate
ROOT into the normal desktop menus.  `xrdcp.1' is a missing man(1)
page.

I've added logic in `configure' to check how libPythia6.so is compiled -
with a single, double, or no trailing underscores.   This seems safer
than to rely on using a specific library, compiled with specific
options.


git-svn-id: http://root.cern.ch/svn/root/trunk@13301 27541ba8-7e3a-0410-8455-c3a389f83636