Minor improvement in TH1::GetAssymetry suggested by Jason Seely Adam Kocoloski pointed out a bug in GetAsymmetry. He suggested a fix, so I'm just forwarding his message on. It's a very minor fix, just need to change the condition in the 'if' statement:
if(bot < 1) to if(bot<1e-6) (or something small)

The problem was that the errors would be computed wrong if there were <1 event
in a bin. Since there is a weighting factor, it's possible to have a
statistically significant number of events in a particular bin, but when the
weight is applied, the value in the bin could be <1. So the errors should be
computed properly in this case, but GetAsymmetry was ignoring this case.


git-svn-id: http://root.cern.ch/svn/root/trunk@13380 27541ba8-7e3a-0410-8455-c3a389f83636