From Wim:
 o) added pythonization of TClonesArray

By enforcing the use of At() (from TObjArray) instead of choosing between
the operator[](int)'s of TClonesArray, __getitem__ will always be safe.
__setitem__ has been reimplemented to allow object assignments, but it has
to be used with a little care: only temporaries will work natural, as it
takes over ownership (feature of the TClonesArray). For example:

 >>> c = TClonesArray( "MyClass" )
 >>> c[ 0 ] = MyClass()
 >>> c[ 1 ] = MyClass()
 >>> m = MyClass()
 >>> c[ 1 ] = m               # safe: will call ~MyClass() on the old one
 >>> c[ 1 ] = MyClass()       # m is now destroyed (None)

Whereas removal works with del, as usual:

 >>> del c[1]

Needless to say that all the above comes with a price and thus it is usually
better to simply copy the contents of a TClonesArray in a list:

 >>> l = list( c )

Modify to your heart's content, then copy *copies* of the elements back to the
TClonesArray for writing out to disk:

 >>> for i in range(len(l)):
 >>>    c[i] = MyClass( l[i] )


git-svn-id: http://root.cern.ch/svn/root/trunk@13640 27541ba8-7e3a-0410-8455-c3a389f83636