From Richard: Fix problem found with TH3 painting via g3d's TMarker3DBox to GL Viewer.
Now GL viewer will take all objects if passed with fID == 0 - i.e. does
not test to see if already cached (as previously). In this case we force
acceptance of all objects offered to viewer as we can't rebuild the
scene.

Updated TMarker3DBox so when recycling object via static method, fID is
zero.


git-svn-id: http://root.cern.ch/svn/root/trunk@13656 27541ba8-7e3a-0410-8455-c3a389f83636