From Axel, Philippe, Stefan: reflex: no shadow classes, data member offset=0 - FIXME no c'tor stub for protected c'tors or abstract no auto-c'tor if protected c'tor func ptr func args cast from void* escape "&" in class names
newlink:
no auto header includes if G__CPPLINK_H is 0 (like for reflex)

I've checked on all root dicts, building the failing ones with -cint;
linking fails on a subset of the dicts that compile - I forgot to remove
some c'tor decls for abstract classes and protected c'tors :-/ I have
not done extensive dict tests. Stefan has acknowledged the func ref
change in the dicts.

rootcint:
add the implementation of the method declared in the ClassDef
and the I/O wrapper and utility code when generating the
dictionary with gccxml.


git-svn-id: http://root.cern.ch/svn/root/trunk@13673 27541ba8-7e3a-0410-8455-c3a389f83636