From Christophe Delaere and Andrea Bocci: Andrea has extended a bit ROOT's TMultiLayerPerceptron to optionally use cross-entropy errors, which allows to train a network for pattern classification based on Bayesian posterior probability. Reference: [Bishop 1995 , Neural Networks for Pattern Recognition], in particular chapter 6.
In order to achieve this, I had to add the softmax (generalized
sigmoid) neuron function, which in turn required a bit of changes to
the neuron itself.
Also, I added softmax and sigmoid as possible output neurons, requiring
some changes to how error back propagation is performed.

Currently, softmax neurons are used only in the output layer, but
everything is setup so that they should be OK as hidden units, too,
provided they form a whole layer.


git-svn-id: http://root.cern.ch/svn/root/trunk@13804 27541ba8-7e3a-0410-8455-c3a389f83636