Improvement of TMath::RootsCubic by Francois-Xavier Gentit
If you call TMath::RootsCubic with the following coefficients:
  coef[3] = 1, coef[2] = -2, coef[1] = 1, coef[0] = -2
the polynomial should have the following roots:
  2 | +i | -i
and the result given by RootsCubic is:
   a b c : 0.666667   2   0.666667     ??????

It is also strange that RootsCubic does not distinguish the 2 cases
 (1) - 3 real roots
 (2) - 1 real root and 2 complex conjugates ones
The 2 cases could be distinguished for instance by returning a boolean?
I join to this mail a possible replacement for RootsCubic correcting these 2
problems.


git-svn-id: http://root.cern.ch/svn/root/trunk@13963 27541ba8-7e3a-0410-8455-c3a389f83636