Add a new bit (kIgnoreTObjectStreamer) which is still BIT(13) for backward compatibility.
Add a new method Clear which support the option "built".  This allows
the un-doing of the Build method (used when loading a library __after__
the creation of an emulated class).

BuildCheck now issues a warning the first time it sees a file
which contains a StreamerInfo which does not match a StreamerInfo
for the same version which is already in memory.

See the test 'warning' in roottest/root/meta/evolution for a sample
of the possible warning scenarii.

Add support for a StreamerInfo in slot '-1' in the case of Foreign
for which the 'current' StreamerInfo is requested after a file
has been read-in (In which case, the TClass version will be -1,
the file StreamerInfo version will be 1, and the in-memory/current
StreamerInfo version will be -1.


git-svn-id: http://root.cern.ch/svn/root/trunk@14332 27541ba8-7e3a-0410-8455-c3a389f83636