Renumber the bits to avoid using BIT(13) which is already used by TObject.  Add to a new bit (kWarned) to avoid duplicating the error messages.
Add a new method PostLoadCheck which is called once the dictionary
for TClass has been completely loaded.  For now this is used to
check the consistency with previously read TStreamerInfos.

When loading the library implementing a class with a ClassDef
__after__ opening a file containing this type of object.
PostLoadCheck will issue a warning if the previously StreamerInfo
does not match the content of the class.  In this case the file
will be readable BUT the writing of object of this class has an
undefined behavior (aka most likely wrong).

Update to GetClassVersion to always return a positive version number.
-1 now indicates a foreign classes for which a non matching
TStreamerInfo has been read before the TStreamerInfo described the
in-memory class has been created.

In conjunction with the TStreamerInfo check-in this also fixes a
problem with the offset of the elements inside TStreamerInfo were
not properly re-initialized when a shared library was loaded after
a file's TStreamerInfo had been used.


git-svn-id: http://root.cern.ch/svn/root/trunk@14333 27541ba8-7e3a-0410-8455-c3a389f83636