From Masa, Axel, Philippe: import of CINT tag v5-16-11 from the CINT CVS. For what else is new see http://root.cern.ch/root/Cint.phtml?relnote.
This includes several parsing improvements (see CINT changelog for
details)

In particular this includes support for distinguish (from the
string based interface) a template of Double32_t from a template
of double.  This allows the storing of the data member as float.

Note:  The compiler still can not distinguish the 2 forms and thus
       only 'string' based interface will be able to handle this
       distinction.


git-svn-id: http://root.cern.ch/svn/root/trunk@14705 27541ba8-7e3a-0410-8455-c3a389f83636